+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `onScroll` and `onScrollBottom` props in chat component
 * Radio group component
 * Possibility to override icon props in `tabs` and `button-tab`
-* Textarea support for background color and horizontal padding as props
+* Textarea support for background color, horizontal padding and font size as props
 * Input keyboard type support
 * Input Form component
 * Date Input support for margin and border overriding as props